MOVEONLY: Move some consensus functions to consensus/consensus.h
Functions definitions are moved to 2 independent files:

To consensus/txverify.cpp:

from main.cpp:
-CheckTransaction
-Consensus::CheckTxInputs
-GetLegacySigOpCount
-GetP2SHSigOpCount
-IsFinalTx

To consensus/blockverify.cpp

from main.cpp:
-CheckBlockHeader
-ContextualCheckBlockHeader
-IsSuperMajority

from pow.cpp:
-CalculateNextWorkRequired
-CheckProofOfWork
-GetNextWorkRequired
